+++ docs/src/literate-gallery/quasi_incompressible_hyperelasticity.jl
@@ -73,7 +73,7 @@
 # ## Implementation
 # We now get to the actual code. First, we import the respective packages
 
-using Ferrite, Tensors, ProgressMeter, WriteVTK
+using Ferrite, Tensors, ProgressMeter, VTKHDF
 using BlockArrays, SparseArrays, LinearAlgebra
 
 # and the corresponding `struct` to store our material properties.
@@ -312,8 +312,8 @@ function solve(interpolation_u, interpolation_p)
     Δt = 0.1
     NEWTON_TOL = 1.0e-8
 
-    pvd = paraview_collection("hyperelasticity_incomp_mixed")
-    for (step, t) in enumerate(0.0:Δt:Tf)
+    vtkhdf = VTKHDFGridFile("hyperelasticity_incomp_mixed.vtkhdf", dh; temporal = true)
+    for t in 0.0:Δt:Tf
         ## Perform Newton iterations
         Ferrite.update!(dbc, t)
         apply!(w, dbc)
@@ -342,12 +342,11 @@ function solve(interpolation_u, interpolation_p)
         end
 
         ## Save the solution fields
-        VTKGridFile("hyperelasticity_incomp_mixed_$step", grid) do vtk
+        write_timestep(vtkhdf, t) do vtk
             write_solution(vtk, dh, w)
-            pvd[t] = vtk
         end
     end
-    vtk_save(pvd)
+    close(vtkhdf)
     vol_def = calculate_volume_deformed_mesh(w, dh, cellvalues)
     print("Deformed volume is $vol_def")
     return vol_def